get-sources.rst: Speed up the repo sync (#108)
By adding --depth=1 and -j 16 the repo sync will be a lot smaller and quicker and in general fit for our purpose.

In case the developer wants to contribute to the repo with some patches or changes, a full copy of the repo might be needed. Therefore added instruction on how to get a full (unshallow) copy.