Bump required kernel version to 3.10
Plasma Mobile for Halium is now based on Ubuntu 18.04, which includes
systemd 237. It should not be compatible with Linux <3.13, but luckily
we do not require all systemd features pegging the version requirement
so high. Testing seems to indicate that Linux 3.10 or higher is
required to boot a Halium OS including systemd 237.

This commit also states the non-Mediatek chipset requirement more
clearly.