Bump required kernel version to 3.10 (#127)
* Bump required kernel version to 3.10

Plasma Mobile for Halium is now based on Ubuntu 18.04, which includes
systemd 237. It should not be compatible with Linux <3.13, but luckily
we do not require all systemd features pegging the version requirement
so high. Testing seems to indicate that Linux 3.10 or higher is
required to boot a Halium OS including systemd 237.

The first thing noted is that kernel and Lineage source must be available.
 I've also changed the list so it's a definition list, which is formatted better
 for our purposes here.